requirements.txt
torch==1.4.0
torchvision==0.5.0
opencv-python==3.4.5.20
PyYAML==6.0
lmdb==1.3.0
Pillow==8.4.0
glob2==0.7
fire==0.4.0